Refactor: definitive code health, architecture, and performance overhaul
- Centralized and deduplicated all core utility functions (strings, security, otp, cookies, ai) into Backend/utils/.
- Adopted a modular backend architecture:
    - Separated CORS and CSRF configurations.
    - Decoupled Socket.io logic into a dedicated service.
- Optimized database performance:
    - Implemented atomic MongoDB operations for real-time messaging and reactions.
    - Added .lean() to performance-critical read queries.
- Enhanced application stability:
    - Fixed critical frontend memory leaks (timers, Speech API).
    - Improved backend logging with a level-based centralized logger.
    - Enhanced database graceful shutdown (SIGTERM support).
- Improved developer experience and security:
    - Enhanced JWT payloads with user IDs.
    - Standardized API response patterns and error handling.
- Fixed UI bugs and optimized React hook dependencies.
